refactor: extract function to diferent files in order t make the code more friendly
